+++ include/iutest_config.hpp
@@ -308,6 +308,10 @@
 #  define IUTEST_HAS_MATCHER_POINTWISE          1
 #endif
 
+#if !defined(IUTEST_HAS_MATCHER_OPTIONAL)
+#  define IUTEST_HAS_MATCHER_OPTIONAL           1
+#endif
+
 #if !defined(IUTEST_USE_THROW_ON_ASSERTION_FAILURE)
 /**
  * @brief   失敗時に例外を throw します。
@@ -805,6 +809,26 @@
 /**
  * @}
 */
+
+/**
+ * @brief   iutest has testsuite api
+*/
+#define IUTEST_HAS_TESTSUITE    1
+
+/**
+ * @brief   iutest has testcase api
+*/
+#if !defined(IUTEST_HAS_TESTCASE)
+#  define IUTEST_HAS_TESTCASE   1
+#endif
+
+#if defined(IUTEST_REMOVE_LEGACY_TEST_CASEAPI_)
+#  undef IUTEST_REMOVE_LEGACY_TEST_CASEAPI_
+#endif
+
+#if !IUTEST_HAS_TESTCASE
+#  define IUTEST_REMOVE_LEGACY_TEST_CASEAPI_
+#endif
 
 /**
  * @private