+++ test/printers_tests.cpp
@@ -577,8 +577,9 @@ IUTEST(PrintToTest, Variant)
         IUTEST_PRINTTOSTRING_EQ(ck, v);
         IUTEST_STREAMOUT_CHECK(v);
     }
-#if IUTEST_HAS_EXCEPTIONS
+#if IUTEST_HAS_EXCEPTIONS && !defined(IUTEST_USE_GTEST)
     {
+        // gtest no support variant valueless
         PrintToLogChecker ck("valueless_by_exception");
         ::std::variant<int, float, AlwaysThrow> v = 0.2f;
         try